+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Upgraded dependencies
 
+### Fixed:
+- Delete category question and confirmation label
+
 ## [0.8.287] - 2023-03-22
 ### Changed:
 - Habit completion card layout